# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /abandon-code:
        get:
            tags:
                - AbandonCodeCURD
            operationId: AbandonCodeCURD_GetAbandonCodeList
            parameters:
                - name: idx1List
                  in: query
                  description: MARK REPLACE IDX COL START 替换内容，索引字段
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
                - name: idx2List
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
                - name: idx3List
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetAbandonCodeListResponse'
        post:
            tags:
                - AbandonCodeCURD
            description: |-
                MARK REPEAT API START 一个表的接口定义
                 --------------------------------------------------
                 tbl : abandon_code
            operationId: AbandonCodeCURD_AddAbandonCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddAbandonCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddAbandonCodeResponse'
        delete:
            tags:
                - AbandonCodeCURD
            operationId: AbandonCodeCURD_DelAbandonCodeByIdx1List
            parameters:
                - name: idx1List
                  in: query
                  description: MARK REPLACE PRIMARY IDX START
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - AbandonCodeCURD
            description: MARK REMOVE IF NO PRIMARY KEY START
            operationId: AbandonCodeCURD_UpdateAbandonCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdateAbandonCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.UpdateAbandonCodeResponse'
...
components:
    schemas:
        api.AbandonCodeInfo:
            type: object
            properties:
                idx1:
                    type: integer
                    description: MARK REPLACE PB COL START 替换内容，所有字段
                    format: int32
                col1:
                    type: string
            description: |-
                --------------------------------------------------
                 tbl : abandon_code
        api.AddAbandonCodeRequest:
            type: object
            properties:
                abandonCode:
                    $ref: '#/components/schemas/api.AbandonCodeInfo'
        api.AddAbandonCodeResponse:
            type: object
            properties:
                abandonCode:
                    $ref: '#/components/schemas/api.AbandonCodeInfo'
...
        api.Empty:
            type: object
            properties: {}
...
        api.GetAbandonCodeListResponse:
            type: object
            properties:
                abandonCodeList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.AbandonCodeInfo'
...
        api.UpdateAbandonCodeRequest:
            type: object
            properties:
                abandonCode:
                    $ref: '#/components/schemas/api.AbandonCodeInfo'
            description: MARK REMOVE IF NO PRIMARY KEY START
        api.UpdateAbandonCodeResponse:
            type: object
            properties:
                abandonCode:
                    $ref: '#/components/schemas/api.AbandonCodeInfo'
...
servers:
    - description: PGO API
      url: http://127.0.0.1:8080